// src/app/models/playing-manager.model.ts
/**
 * Playing管理 IF
 */
export interface PlayingManager {
  // 表示回数 Top
  displayedTop: number;
  // 表示回数 Web
  displayedWeb: number;
  // 表示回数 Web
  displayedSoccer: number;
  // 表示回数 Big Data
  displayedBigData: number;
  // 表示回数 Coding
  displayedCoding: number;
  // 表示回数 Math
  displayedMath: number;
  // 表示回数 Economy
  displayedEconomy: number;
  // 表示回数 Ai
  displayedAi: number;
}